Stock price change from INTEGER to NUMERIC

+++ migrations/20150813205109-create-stocks.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.NUMERIC
       }
     });
   },
